x/toe/types/messages_topic_test.go
package types

import (
	"testing"

	"github.com/WilliamKelly00/toev2/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateTopic_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateTopic
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateTopic{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateTopic{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateTopic_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateTopic
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateTopic{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateTopic{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteTopic_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteTopic
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteTopic{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteTopic{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
